1. Фильтры для объявлений машин [#4745 #551382]

+++ src/main/java/ru/job4j/cars/repository/EngineRepository.java
@@ -10,9 +10,9 @@ public interface EngineRepository {
 
     void update(Engine engine);
 
-    void deleteById(int id);
+    void deleteById(Long id);
 
-    Optional<Engine> findById(int id);
+    Optional<Engine> findById(Long id);
 
     List<Engine> findAll();
 }
