+++ src/main/java/ru/job4j/cars/service/SimpleUserService.java
@@ -24,7 +24,7 @@ public class SimpleUserService implements UserRepository {
     }
 
     @Override
-    public void deleteById(int id) {
+    public void deleteById(Long id) {
         userRepository.deleteById(id);
     }
 
@@ -39,7 +39,7 @@ public class SimpleUserService implements UserRepository {
     }
 
     @Override
-    public Optional<User> findById(int id) {
+    public Optional<User> findById(Long id) {
         return userRepository.findById(id);
     }
